Added Features
Added More of the map, made actual room collisions work.  Added a sign tile with transparency to the tileset.

TODO: I still need to get the corners working well. I think it might have to do with the collision engine, or maybe the collisions that I have set in the tileset.