Updated Scrolling Class for Teleporting
I moved all of the functions of the scrolling class to one update function. I also added scrolling.change_viewport_margins, which is handy for centering the view on the player. After the viewport updates at the end of scrolling.update(), the viewport is changed to normal.
TODO: Add constant for the regular viewport size to make viewport more modular and easily adjustable.